Move pint/_version.py -> pint/extern/_version.py
