+++ pint/__init__.py
@@ -3,7 +3,7 @@
 PINT Is Not TEMPO3!
 """
 
-from ._version import get_versions
+from .extern._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
