+++ setup.cfg
@@ -6,8 +6,7 @@
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = pint/_version.py
-versionfile_build = pint/_version.py
+versionfile_source = pint/extern/_version.py
+versionfile_build = pint/extern/_version.py
 tag_prefix = ''
 parentdir_prefix = 'pint-'
-
